backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
amadeus==9.0.0
google-genai==0.6.0
motor==3.6.0
requests==2.31.0
PyJWT==2.8.0
